langchain-server1/requirements.txt
langchain
langchain-openai
langchain-community
openai==1.10.0
faiss-cpu==1.11.0.post1
pypdf==4.0.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.10.0
elasticsearch==8.11.0